Change AbstractChangeNamespaceToMatchFolderCodeFixProvider to use the SyncNamespaceDocumentAction instead of the public API
