+++ src/Workspaces/Core/Portable/Rename/Renamer.RenameDocumentAction.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using System.Collections.Immutable;
 using System.Globalization;
 using System.Threading;
@@ -19,41 +20,28 @@ namespace Microsoft.CodeAnalysis.Rename
         /// 
         /// See <see cref="RenameDocumentActionSet" /> on use case and how to apply them to a solution.
         /// </summary>
-        public abstract class RenameDocumentAction
+        public sealed class RenameDocumentAction
         {
-            private readonly ImmutableArray<ErrorResource> _errorStringKeys;
+            private readonly InternalRenameDocumentAction _internalRenameDocumentAction;
 
-            internal RenameDocumentAction(ImmutableArray<ErrorResource> errors)
+            internal RenameDocumentAction(InternalRenameDocumentAction internalRenameDocumentAction)
             {
-                _errorStringKeys = errors;
+                _internalRenameDocumentAction = internalRenameDocumentAction;
             }
 
             /// <summary>
             /// Get any errors that have been noted for this action before it is applied.
             /// Can be used to present to a user.
             /// </summary>
-            public ImmutableArray<string> GetErrors(CultureInfo? culture = null)
-                => _errorStringKeys.SelectAsArray(s => string.Format(WorkspacesResources.ResourceManager.GetString(s.FormatString, culture ?? WorkspacesResources.Culture)!, s.Arguments));
+            public ImmutableArray<string> GetErrors(CultureInfo? culture = null) => _internalRenameDocumentAction.GetErrors(culture);
 
             /// <summary>
             /// Gets the description of the action. Can be used to present to a user to describe
             /// what extra actions will be taken.
             /// </summary>
-            public abstract string GetDescription(CultureInfo? culture = null);
+            public string GetDescription(CultureInfo? culture = null) => _internalRenameDocumentAction.GetDescription(culture);
 
-            internal abstract Task<Solution> GetModifiedSolutionAsync(Document document, OptionSet optionSet, CancellationToken cancellationToken);
-
-            internal readonly struct ErrorResource
-            {
-                public string FormatString { get; }
-                public object[] Arguments { get; }
-
-                public ErrorResource(string formatString, object[] arguments)
-                {
-                    FormatString = formatString;
-                    Arguments = arguments;
-                }
-            }
+            internal Task<Solution> GetModifiedSolutionAsync(Document document, OptionSet optionSet, CancellationToken cancellationToken) => _internalRenameDocumentAction.GetModifiedSolutionAsync(document, optionSet, cancellationToken);
         }
     }
 }
