+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/Core/Rename/IChangeNamespaceService.cs
