+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/Core/Rename/InternalRenameDocumentAction.cs
@@ -1,0 +1,19 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Collections.Immutable;
+using System.Globalization;
+using System.Threading;
+using System.Threading.Tasks;
+using Microsoft.CodeAnalysis.Options;
+
+namespace Microsoft.CodeAnalysis.Rename
+{
+    internal abstract class InternalRenameDocumentAction
+    {
+        internal abstract ImmutableArray<string> GetErrors(CultureInfo? culture);
+        internal abstract string GetDescription(CultureInfo? culture);
+        internal abstract Task<Solution> GetModifiedSolutionAsync(Document document, OptionSet _, CancellationToken cancellationToken);
+    }
+}
